+++ app/Filament/Server/Resources/Subusers/Pages/ListSubusers.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Filament\Server\Resources\Users\Pages;
+namespace App\Filament\Server\Resources\Subusers\Pages;
 
-use App\Filament\Server\Resources\Users\UserResource;
+use App\Filament\Server\Resources\Subusers\SubuserResource;
 use App\Traits\Filament\CanCustomizeHeaderActions;
 use App\Traits\Filament\CanCustomizeHeaderWidgets;
 use Filament\Actions\Action;
@@ -10,12 +10,12 @@ use Filament\Actions\ActionGroup;
 use Filament\Resources\Pages\ListRecords;
 use Illuminate\Contracts\Support\Htmlable;
 
-class ListUsers extends ListRecords
+class ListSubusers extends ListRecords
 {
     use CanCustomizeHeaderActions;
     use CanCustomizeHeaderWidgets;
 
-    protected static string $resource = UserResource::class;
+    protected static string $resource = SubuserResource::class;
 
     /** @return array<Action|ActionGroup> */
     protected function getDefaultHeaderActions(): array
