--- app/Policies/Admin/ServerPolicy.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Policies\Admin;
-
-class ServerPolicy
-{
-    use DefaultPolicies;
-
-    protected string $modelName = 'server';
-}
