+++ app/Policies/AllocationPolicy.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Policies\Server;
+namespace App\Policies;
 
 use App\Models\Permission;
 use App\Models\User;
