+++ app/Policies/WebhookConfigurationPolicy.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Policies\Admin;
+namespace App\Policies;
 
 class WebhookConfigurationPolicy
 {
-    use DefaultPolicies;
+    use DefaultAdminPolicies;
 
     protected string $modelName = 'webhook';
 }
